Complete simplification: remove all backwards compatibility and executor abstraction
- Removed optional GraphQL executor and configuration methods
- Eliminated executor module entirely - no longer needed
- Moved GraphQL execution directly into LinearClient as execute_graphql() method
- All SDK methods now use clean GraphQL abstraction by default
- No more fallback logic or backwards compatibility overhead
- Simplified imports and dependencies
- All tests pass, CLI functionality unchanged
- Much cleaner, focused implementation
